Bug #52043: s4.dc._nano2s must return int

+++ services/univention-s4-connector/modules/univention/s4connector/s4/dc.py
@@ -74,7 +74,7 @@ def _s2nano(seconds):
 
 
 def _nano2s(nanoseconds):
-	return nanoseconds / 10000000
+	return int(nanoseconds / 10000000)
 
 
 def ucs2con(s4connector, key, object):
